Social Media Icons update + Dog Image update
Social Media Icons are now redirecting to the Via-Bicho Facebook and Instagram.The dog picture on About Us section is now on the center of the screen in all devices.
