+++ via-bicho/src/Components/DogImageAboutUs/index.js
@@ -1,9 +1,10 @@
 import React from "react";
-import Image from "../../../static/images/dog_about_us.png"
+import Image from "../../assets/images/cute_dog.png"
 
 function DogImageAboutUs() {
+
     return (
-            <img src={Image}/>
+        <img src={Image} alt="dog" />
     )
 }
 export default DogImageAboutUs;
