+++ via-bicho/src/Components/SocialMedia/index.js
@@ -1,15 +1,15 @@
 import React from "react";
-import FacebookIcon from "../../Icons/Facebook.js";
-import InstagramIcon from "../../Icons/Instagram.js";
-import { Container, Logos } from "./styles";
+import FacebookIcon from "../../assets/icons/Facebook.js";
+import InstagramIcon from "../../assets/icons/Instagram.js";
+import { Container, Logos, Cursor} from "./styles";
 
 function SocialMedia() {
     return (
         <Container>
             <h2>{"Redes sociais"}</h2>
-            <Logos>
-                <FacebookIcon width="100px" heigth="100px" />
-                <InstagramIcon width="100px" heigth="100px" />
+            <Logos> 
+                <Cursor><FacebookIcon/></Cursor>
+                <Cursor><InstagramIcon width="100px" heigth="100px" /></Cursor>
             </Logos>
         </Container>
     )
