+++ via-bicho/src/Components/SocialMedia/styles.js
@@ -18,3 +18,7 @@ export const Logos = styled.div`
         justify-content: space-evenly;
         margin-bottom: 40px;
 `;
+
+export const Cursor = styled.div`
+        cursor:pointer;
+`;
