+++ via-bicho/src/assets/icons/index.js
